moved components into their own folder. updated paths
