+++ web/src/components/Errors/Errors.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import GraphQLErrorList from './GraphqlErrorList'
-import Layout from '../Layout'
+import Layout from '../Layout/Layout'
 
 const Errors = ({ errors }) => (
   <Layout>
