+++ web/src/components/Layout/Layout.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Aside from './Aside/Aside'
+import Aside from '../Aside/Aside'
 
 const Layout = ({ children }) => (
   <div className="mx-4 flex flex-col scroll-smooth lg:mx-auto lg:flex-row">
