+++ web/src/pages/404.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Link } from 'gatsby'
-import Layout from '../components/Layout'
+import Layout from '../components/Layout/Layout'
 import { useSiteMetadata } from '../components/Hooks/SiteMeta'
 
 const NotFoundPage = () => {
