+++ web/src/pages/confirmation.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/Layout'
+import Layout from '../components/Layout/Layout'
 import { Link } from 'gatsby'
 
 import { useSiteMetadata } from '../components/Hooks/SiteMeta'
